refactor layouts and add new condition-layout
