+++ pages/auth/recovery-code-failed/index.tsx
@@ -1,10 +1,10 @@
 import React from 'react';
 
 import { RecoveryCodeFailed } from '@/pagesComponents/recoveryCodeFailed/ui/RecoveryCodeFailed';
-import { getAuthLayout } from '@/shared/providers';
+import { getNonAuthorizedLayout } from '@/shared/providers';
 
 export default function Page() {
-  return getAuthLayout(
+  return getNonAuthorizedLayout(
     <>
       <RecoveryCodeFailed />
     </>
