+++ pages/auth/sign-in/index.tsx
@@ -1,7 +1,7 @@
 import React from 'react';
 
 import { SignIn } from '@/pagesComponents';
-import { getAuthLayout } from '@/shared/providers';
+import { getNonAuthorizedLayout } from '@/shared/providers';
 
 export default function Page() {
   return (
@@ -11,4 +11,4 @@ export default function Page() {
   );
 }
 
-Page.getLayout = getAuthLayout;
+Page.getLayout = getNonAuthorizedLayout;
