+++ pages/profile/[id]/index.tsx
@@ -4,7 +4,7 @@ import { wrapper } from '@/myApp/store';
 import { Profile } from '@/pagesComponents/profile/Profile';
 import { getPublicPostsUser, getPublicUserProfile } from '@/shared/api/public/publicApi';
 import { GetPostsResponse, GetPublicUserProfileResponse } from '@/shared/api/public/publicTypes';
-import { getBaseLayout } from '@/shared/providers';
+import { getAuthorizedLayout } from '@/shared/providers';
 import { GetServerSideProps } from 'next';
 
 interface Props {
@@ -47,4 +47,4 @@ export const getServerSideProps: GetServerSideProps = wrapper.getServerSideProps
   };
 });
 
-Page.getLayout = getBaseLayout;
+Page.getLayout = getAuthorizedLayout;
