+++ src/shared/providers/authorizedLayout/AuthorizedLayout.module.scss
