+++ src/shared/providers/index.ts
@@ -1,2 +1,3 @@
-export { AuthLayout, getAuthLayout } from './authLayout/AuthLayout';
-export { BaseLayout, getBaseLayout } from './baseLayout/BaseLayout';
+export { AuthorizedLayout, getAuthorizedLayout } from '@/shared/providers/authorizedLayout/AuthorizedLayout';
+export { ConditionLayout, getConditionLayout } from '@/shared/providers/conditionLayout/ConditionLayout';
+export { NonAuthorizedLayout, getNonAuthorizedLayout } from '@/shared/providers/nonAuthorizedLayout/NonAuthorizedLayout';
