+++ src/shared/providers/nonAuthorizedLayout/NonAuthorizedLayout.module.scss
